Aggregated reports tests into a single test file
